# Backend/api-gateway-service/docs/swagger.yaml
definitions:
...
  responses.DeleteFavouriteResponse:
    properties:
      success:
        type: boolean
    type: object
...
  responses.DeleteMemoryResponse:
    properties:
      success:
        type: boolean
    type: object
...
  responses.DetailedMemoryResponse:
    properties:
      isFavourite:
        type: boolean
      memory:
        $ref: '#/definitions/responses.MemoryResponse'
      tags:
        items:
          type: string
        type: array
    type: object
  responses.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  responses.FavouriteResponse:
    properties:
      memoryId:
        type: integer
      userId:
        type: integer
    type: object
...
  responses.ListMemoriesResponse:
    properties:
      memories:
        items:
          $ref: '#/definitions/responses.DetailedMemoryResponse'
        type: array
    type: object
...
  responses.MemoryResponse:
    properties:
      createdAt:
        type: string
      duration:
        type: number
      groupId:
        type: integer
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      mediaType:
        type: string
      mediaUrl:
        type: string
      musicId:
        type: string
      offset:
        type: number
      updatedAt:
        type: string
      userId:
        type: integer
      userImage:
        type: string
      username:
        type: string
    type: object
...
host: rewindapp.ru
info:
  contact: {}
  description: This is the API Gateway for the Rewind application.
  title: Rewind API Gateway
  version: "1.0"
paths:
...
  /api/groups/{groupId}/memories:
    get:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully listed memories
          schema:
            $ref: '#/definitions/responses.ListMemoriesResponse'
        "400":
          description: Bad Request - Invalid group ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden - Permission denied by service
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found - Group or User not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List memories by group
      tags:
      - memories
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: integer
      - description: Media Type (image/video/quote)
        in: formData
        name: mediaType
        required: true
        type: string
      - description: Media file to upload
        in: formData
        name: mediaFile
        required: true
        type: file
      - description: Latitude (optional)
        in: formData
        name: latitude
        type: number
      - description: Longitude (optional)
        in: formData
        name: longitude
        type: number
      - description: Music ID (optional)
        in: formData
        name: musicId
        type: string
      - description: Offset (optional)
        in: formData
        name: offset
        type: number
      - description: Duration (optional)
        in: formData
        name: duration
        type: number
      - description: Comma-separated tags (optional)
        in: formData
        name: tags
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created memory
          schema:
            $ref: '#/definitions/responses.MemoryResponse'
        "400":
          description: Bad Request - Invalid request data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found - Group or User not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new memory
      tags:
      - memories
  /api/groups/{groupId}/memories/{memoryId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: integer
      - description: Memory ID
        in: path
        name: memoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted memory
          schema:
            $ref: '#/definitions/responses.DeleteMemoryResponse'
        "400":
          description: Bad Request - Invalid memory ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden - Permission denied by service
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found - Memory not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a memory by ID
      tags:
      - memories
    get:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: integer
      - description: Memory ID
        in: path
        name: memoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully got memory
          schema:
            $ref: '#/definitions/responses.DetailedMemoryResponse'
        "400":
          description: Bad Request - Invalid memory ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden - Permission denied by service
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found - Memory not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a memory by ID
      tags:
      - memories
  /api/groups/{groupId}/memories/{memoryId}/favourite:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: integer
      - description: Memory ID
        in: path
        name: memoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully unmarked as favourite
          schema:
            $ref: '#/definitions/responses.DeleteFavouriteResponse'
        "400":
          description: Bad Request - Invalid memory ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden - Permission denied by service
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found - Memory or User not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Unmark a memory as favourite
      tags:
      - memories
    post:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: integer
      - description: Memory ID
        in: path
        name: memoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully marked as favourite
          schema:
            $ref: '#/definitions/responses.FavouriteResponse'
        "400":
          description: Bad Request - Invalid memory ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden - Permission denied by service
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found - Memory or User not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Conflict - Memory is already a favourite
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Mark a memory as favourite
      tags:
      - memories
...
securityDefinitions:
  ApiKeyAuth:
    description: Enter your access token in the format "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"